move now deprecated and unused extension function to OnionUrl companion
